@@ -37,8 +37,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
 		<version>2.3.2</version>
         <configuration>
-          <source>1.7</source>
-          <target>1.7</target>
+          <source>15</source>
+          <target>15</target>
           <debug>true</debug>
           <debuglevel>lines,source</debuglevel>
         </configuration>
@@ -88,12 +88,6 @@
 	    		<artifactId>maven-jar-plugin</artifactId>
 	    		<version>2.4</version>
 	    	</plugin>
-                <plugin>
-                    <groupId>org.apache.felix</groupId>
-                    <artifactId>maven-bundle-plugin</artifactId>
-                    <version>3.2.0</version>
-                    <extensions>true</extensions>
-                </plugin>
     	</plugins>
     </pluginManagement>
   </build>
@@ -181,6 +175,6 @@
                 <url>http://bits.netbeans.org/maven2/</url>
             </repository>
         </repositories>
-        
+
 
 </project>
